SensorDataService methods updated with Exceptions.

+++ BusinessLogicLayer/SensorDataService.cs
@@ -29,7 +29,7 @@ namespace BusinessLogicLayer
             {
                 if (sensorData.SensorId == null)
                 {
-                    throw new ArgumentNullException(nameof(sensorData.SensorId));
+                    throw new Exception("The received SensorData cannot be stored in the database, because it does not contain SensorId.");
                 }
                 else
                 {
@@ -37,7 +37,7 @@ namespace BusinessLogicLayer
 
                     if (sensor == null)
                     {
-                        throw new ArgumentNullException(nameof(sensor));
+                        throw new Exception($"The received SensorData contains {sensorData.SensorId} SensorId, that does not correspond to any sensor in the database.");
                     }
                     else
                     {
@@ -50,7 +50,30 @@ namespace BusinessLogicLayer
         // Get SensorData of the specific sensor.
         public async Task<List<SensorData>> GetSensorsDataBySensorIdAsync(string sensorId)
         {
-            return await _dataService.GetSensorsDataBySensorIdAsync(sensorId);
+            try
+            {
+                if (string.IsNullOrWhiteSpace(sensorId))
+                {
+                    throw new ArgumentNullException(nameof (sensorId));
+                }
+                else
+                {
+                    var result = await _dataService.GetSensorsDataBySensorIdAsync(sensorId);
+
+                    if (result == null)
+                    {
+                        throw new Exception($"The list of SensorData, that correspond to the sensor with the {sensorId} Id, is not found.");
+                    }
+                    else
+                    {
+                        return result;
+                    }
+                }
+            }
+            catch
+            {
+                throw;
+            }
         }
 
         // Get SensorData of the specific sensor with pagination.
@@ -58,13 +81,67 @@ namespace BusinessLogicLayer
                                                                                                                   int page,
                                                                                                                   int pageSize)
         {
-            return await _dataService.GetSensorsDataPaginationAsync(sensorId, page, pageSize);
+            try
+            {
+                if (string.IsNullOrWhiteSpace(sensorId))
+                {
+                    throw new ArgumentNullException(nameof (sensorId));
+                }                
+                else if (page <= 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof (page));
+                }                
+                else if (pageSize <= 0)
+                {
+                    throw new ArgumentOutOfRangeException(nameof (pageSize), "The page size must be greater than 0.");
+                }
+                else
+                {
+                    var (totalPages, readOnlyList) = await _dataService.GetSensorsDataPaginationAsync(sensorId, page, pageSize);
+
+                    if (readOnlyList == null)
+                    {
+                        throw new Exception("The list of SensorData is not found.");
+                    }
+                    else
+                    {
+                        return (totalPages, readOnlyList);
+                    }
+                }
+            }
+            catch
+            {
+                throw;
+            }
         }
 
         // Delete all SensorData of the specific sensor.
         public async Task<DeleteResult> DeleteAllSensorDataBySensorIdAsync(string sensorId)
         {
-            return await _dataService.DeleteAllSensorDataBySensorIdAsync(sensorId);
+            try
+            {
+                if (string.IsNullOrEmpty(sensorId))
+                {
+                    throw new ArgumentNullException(nameof (sensorId));
+                }
+                else
+                {
+                    var result = await _dataService.DeleteAllSensorDataBySensorIdAsync(sensorId);
+
+                    if (result == null)
+                    {
+                        throw new Exception("Could not delete SensorData.");
+                    }
+                    else
+                    {
+                        return result;
+                    }
+                }
+            }
+            catch
+            {
+                throw;
+            }
         }
     }
 }
